Refactored the Thread component, divided it into smaller components. Ready to pull the Dev branch.
