+++ frontend/src/components/Thread/Thread.js
@@ -3,12 +3,16 @@ import { useDispatch, useSelector } from 'react-redux'
 import { Buffer } from 'buffer'
 import { Avatar, Button, Card } from 'antd'
 import { useNavigate, useParams } from 'react-router-dom'
-import { deleteThreadAsync, getThreadAsync } from '../thunk/threadThunk'
-import '../style/Thread.css'
-import useAxiosPrivate from '../hooks/useAxiosPrivate.js'
+import { deleteThreadAsync, getThreadAsync } from '../../thunk/threadThunk'
+import '../../style/Thread/Thread.css'
+import useAxiosPrivate from '../../hooks/useAxiosPrivate.js'
 import axios from 'axios'
 import { format } from 'date-fns'
-import UpdateThreadForm from './UpdateThread/UpdateThreadForm'
+import UpdateThreadForm from '../UpdateThread/UpdateThreadForm'
+import UserInfo from './UserInfo';
+import ThreadTitle from './ThreadTitle';
+import PetCard from './PetCard';
+import ThreadContent from './ThreadContent';
 
 const { Meta } = Card;
 
@@ -96,44 +100,58 @@ function Thread() {
 
   const {title, content} = thread;
 
+  // return (
+  //   <div className="thread-container">
+  //     <div className="user-info">
+  //       <Avatar size={64} src={poster.avatar} alt="user avatar" />
+  //       <h2 className="thread-username">{poster.username}</h2>
+  //       {user && user.id === poster._id && (
+  //         <div className="button-wrapper">
+  //           <Button className="edit-button" onClick={handleEdit}>Edit Thread</Button>
+  //           <Button className="delete-button" onClick={handleDelete}>Delete Thread</Button>
+  //         </div>
+  //       )}
+  //     </div>
+  //     <div className="thread-title-container">
+  //       <h3 className="thread-title">{title}</h3>
+  //     </div>
+  //     <Card
+  //       className="id-card"
+  //       cover={<img className="id-card-img" alt="pet" src={getPetPicUrl()} />}
+  //     >
+  //       <Meta
+  //         title={<span className="id-card-title">Name: {pet.name}</span>}
+  //         description={
+  //           <div className="id-card-info">
+  //             <p><span className="id-card-label">ID: </span>{pet.id}</p>
+  //             {pet.species === 'Cat' ?
+  //               <p><span className="id-card-label">Cat Breed: </span>{pet.breed}</p> :
+  //               <p><span className="id-card-label">Dog Breed: </span>{pet.breed}</p>
+  //             }
+  //             <p><span className="id-card-label">Sex: </span>{pet.sex}</p>
+  //             <p><span className="id-card-label">Last Seen Time: </span>{format(new Date(pet.lastSeenTime), 'hh:mm aa, MMMM do yyyy')}</p>
+  //             <p><span className="id-card-label">Description: </span>{pet.description}</p>
+  //           </div>
+  //         }
+  //       />
+  //     </Card>
+  //     <div className="thread-text-container">
+  //       <p className="thread-text">{content}</p>
+  //     </div>
+  //     <UpdateThreadForm open={editModalVisible}
+  //                       onUpdate={handleUpdate}
+  //                       onCancel={handleCancel}
+  //                       threadId={id}
+  //     />
+  //   </div>
+  // );
+
   return (
     <div className="thread-container">
-      <div className="user-info">
-        <Avatar size={64} src={poster.avatar} alt="user avatar" />
-        <h2 className="thread-username">{poster.username}</h2>
-        {user && user.id === poster._id && (
-          <div className="button-wrapper">
-            <Button className="edit-button" onClick={handleEdit}>Edit Thread</Button>
-            <Button className="delete-button" onClick={handleDelete}>Delete Thread</Button>
-          </div>
-        )}
-      </div>
-      <div className="thread-title-container">
-        <h3 className="thread-title">{title}</h3>
-      </div>
-      <Card
-        className="id-card"
-        cover={<img className="id-card-img" alt="pet" src={getPetPicUrl()} />}
-      >
-        <Meta
-          title={<span className="id-card-title">Name: {pet.name}</span>}
-          description={
-            <div className="id-card-info">
-              <p><span className="id-card-label">ID: </span>{pet.id}</p>
-              {pet.species === 'Cat' ?
-                <p><span className="id-card-label">Cat Breed: </span>{pet.breed}</p> :
-                <p><span className="id-card-label">Dog Breed: </span>{pet.breed}</p>
-              }
-              <p><span className="id-card-label">Sex: </span>{pet.sex}</p>
-              <p><span className="id-card-label">Last Seen Time: </span>{format(new Date(pet.lastSeenTime), 'hh:mm aa, MMMM do yyyy')}</p>
-              <p><span className="id-card-label">Description: </span>{pet.description}</p>
-            </div>
-          }
-        />
-      </Card>
-      <div className="thread-text-container">
-        <p className="thread-text">{content}</p>
-      </div>
+      <UserInfo poster={poster} user={user} handleEdit={handleEdit} handleDelete={handleDelete} />
+      <ThreadTitle title={title} />
+      <PetCard pet={pet} getPetPicUrl={getPetPicUrl} />
+      <ThreadContent content={content} />
       <UpdateThreadForm open={editModalVisible}
                         onUpdate={handleUpdate}
                         onCancel={handleCancel}
