+++ frontend/src/components/Thread/ThreadContent.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import '../../style/Thread/ThreadContent.css'
+
+const ThreadContent = ({ content }) => (
+  <div className="thread-text-container">
+    <p className="thread-text">{content}</p>
+  </div>
+);
+
+export default ThreadContent;
